requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.12.0
oauth2client==4.1.3
anthropic==0.7.8
python-dotenv==1.0.0
num2words==0.5.12
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0 